Fixed update of stream closed status in case of data frames.
It was possible that a client sent a data frame with close=true, but the half close of the
stream was delayed until after all the data bytes were sent.
The server received the data bytes and sent a reply with close=true that could have been
processed before the stream was actually half closed by the data frame delayed update,
and this generated a race condition.
Now the close status is updated before sending the data (like all other frames), but the
stream is removed only when data has been fully sent.