Fixed spin loop caused by interaction of HttpGenerator._bypass and HttpGenerator.isBufferFull().
In HttpOutput.write(), an infinite spin loop was triggered when a previous write() set _bypass to true,
but _bypass was not reset after the write was complete. An additional write would have caused
a call to isBufferFull(), implemented in terms of _bypass, which would have returned true erroneously,
and this would have caused the infinite spin loop.
Now we properly reset _bypass when we fully write a data frame, so that isBufferFull() can properly
return false and allow for further writes.