Jetty9 - Better implementation for onOpen().
Removed the set of the clientMode, since this needs to be set by the creator of the SslConnection,
depending whether is a client or a server.
Also, always calling super.onOpen() to call fillInterested(), even on client side, because it will need
to read the server SSL handshake.
Finally, removed the delegated call to onOpen() to the application connection. This is already done
by the creator of the application connection.