400849 - Conversation hangs if non-first request fails when queued.
Taken the chance to revisit the lifecycle of HttpConversation and HttpExchange
(now they are created as soon as the request is sent, before they were created
when the exchange was associated to the connection).
This change made the conversation listeners available as soon as the request
was sent, so that it was simpler to notify the right listeners also in case of failure
of a non-first request (which fixes the bug).