376273 - Early EOF because of SSL Protocol Error on https://api-3t.paypal.com/nvp.
This was caused by the fact that the other peer closed the raw socket after sending data.
SslConnection was reading the data, but not notifying the application of the data, then
reading the abrupt close, which was causing an exception, ending up in the application
never being notified of the data that arrived.
Now we catch and ignore the exception during SSLEngine.closeInbound(), and we properly
send an alert to the other peer (instead of hard closing the connection as well).
In this way, the application has the chance to read the data and then close the connection.