409028 - Jetty HttpClient does not work with proxy CONNECT method.
The problem was due to the fact that the server replied with HTTP/1.0
to the CONNECT request; because of this, the parser was set as non
persistent, and the tunnel was immediately closed.

Now we are setting the parser as persistent if the method is a CONNECT,
no matter what HTTP version the server specifies.