Close JDBC statements in JDBCSessionManager and JDBCSessionIdManager.
Not closing JDBC Statements and PreparedStatements lead to cursor leakages in
the database.

ResultSets are not explicitly closed as they are closed automatically when
their statement is closed.

The bug was discovered when using Oracle OJDBC6 11.2.0.3 with the Oracle
Universal Connection Pool (UCP) 11.1.0.7.

Calling Connection.close() on a borrowed connection does not cause the
connections' statements to be closed.