450855 GzipFilter MIGHT_COMPRESS exception
Disabled tests that tried to apply the GzipFilter to default async content generation.  These
tests will never work as GzipFilter wraps the response and the default async does not preserve wrappers.

For the MIGHT_COMPRESS issue, improved the filter to handle the case that it has already been applied to
a request/response and that the state is already mightCompress.