Various fixed to make WebSocket work in Android 5.
The main issue is that after a wrap/unwrap operation, the Android
SSLEngine implementation does not return the right number of bytes
produced/consumed in SSLEngineResult.
The SslConnection implementation was relying on these values to
properly update the Jetty Buffer indexes.
The fix consist in not using the SSLEngineResult values to update the
Jetty Buffer indexes, but using the underlying ByteBuffer index
changes to update the Jetty Buffer indexes.