Rewrite of ManagedSelector to handle connection creation as an ExecutionStrategy task.
Now the creation of a connection, and the Connection.onOpen() call
happen as a Runnable that is run by the ExecutionStrategy.
This allows onOpen() to block or otherwise perform tasks that are not
suitable to be run by a selector thread, since the ExecutionStrategy
will guarantee that another thread will take over the selector duties.