Closing the connection also in case the parser is in CLOSED state.
When the request is fully read and the response is sent, the parser
may enter the CLOSE state if there were Connection: close headers.

It may happen that later the connection reads -1, then moves the
parser is CLOSED state, so the condition to close the connection must
include the CLOSED state.