Moved console-capture.xml from modules/console-capture/ to etc/.
This is needed to ease migrations from Jetty 9.3.x, that otherwise
won't be able to find that file.
Furthermore, it's a file that will be rarely modified, since it can
be tuned via properties in start.ini.
Finally, it needs to be in etc/ if it is to be referenced by jetty.conf.