Completed speeding up Multipart for large parts.
In essence we speed up multipart requests by reducing the amount of branching done. This is done by reading chunks (byte arrays) of data at once. Although this does result in faster multi part parsing (1.3s to 0.3s for about 3k requests) it does lead to some sticky situations where we need to undo how much we have read. To do this I have used mark/reset however for base64 encoding this is not possible and instead careful work had to be done to ensure each read would return a very specific amount of data. Also in this implementation we can write one char to much which may result in a file being created to early (e.g. we are supposed to create a file at 50bytes but we create it at 49), as we write one extra char we have to do some special work to backup in the case we have written to a file we truncate the file otherwise we backup the count into the byte array output stream.