Issue #1363
This is just some renaming withing HttpInput to improve it's understandability, plus a small
amount of refactoring for EOF handling.

This does not yet fix the deadlock, but is working towards better EOF handling so that state
can be pushed from HttpInput to HttpChannelState and avoid the need to query HttpInput state.