Fixed determining mime type of `PathResource` based on its file name. (#1532)
The result of `toString()` is expected to be the path name of a Resource. This is used for determining the mime type based on the file name extension in `org.eclipse.jetty.server.ResourceCache.Content`. Without a custom `toString()` implementation, no mime type can be determined for a resource. This problem does no longer exist in current Jetty, since `PathResource` has vanished.