Fixes #1718 - QueuedThreadPool not exposed on JMX.
Now MBeanContainer keeps track of the parent/child component relationship.

MBeans are registered for a pair (parent, child) and unregistered only
for the same pair.

For example, Server owns a ThreadPool and a Connector; in turn the
Connector adds the Server ThreadPool as an unmanaged component to itself.
The ThreadPool is registered as MBean with the pair (Server, ThreadPool).
When the Connector is stopped and removed from the component tree,
MBeanContainer sees a removal event for the pair (ServerConnector, ThreadPool).
But since that pair was not the one that triggered the registration of
the MBean, the MBean is not unregistered.