Improved ManagedSelector
The actions should not be none blocking and they are intended to be executed
by the selector thread between selections, so instead of passing them back
to an execution strategy that might dispatch them, execute them in a tight
loop with lock held.  If a blocking task is discovered, warn and then pass
back to the exection strategy.

Signed-off-by: Greg Wilkins <gregw@webtide.com>