Issue #2349 - Review HTTP/2 max streams enforcement.
Changed the max concurrent streams enforcement to be based on halves.

We now allow 2*max_concurrent_streams+1 halves to cope with the
race condition where the server sent the last frame of a stream,
but before it notifies the callback a new (legit) request arrives
from the client.

The new logic is not perfect as in the worst case it allows more
streams than max_concurrent_streams, but has an upper bound and
the window of the worst case happening is very small, especially
for non-loopback networks.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>