Fixes #2366 - Review HTTP/2 interleaving.
Now using a fairer algorithm that performs multiple generation passes
but without generating all frames for the current entries - it stops
after N passes over all the entries when the number of bytes to
write is large enough.

Using HttpConfiguration.outputBufferSize as the writeThreshold to
stop the generation after a pass over all the entries.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>