Issue #2998 - Cleanup the dump implementation.
Fixed dump() in QueuedThreadPool.
When the dump was not detailed, it was printing jobs=0 even if there
were jobs in the queue.
Given that it was adding no information (actually misleading information)
and that the queue size is already reported by QueuedThreadPool.toString()
the jobs are not dumped if the dump is not detailed.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>