Made Utf8Encoders tolerant to incomplete sequences.
The idea of this patch is preventing throw an exception and fail the request if
somewhere in beggining or the middle of a form parameter appears an incomplete
sequences of UTF-8 chars.

Right now the code tolerance this only at the end, but it is definitely possible
to have this at the middle the POST parameter or GET string if a remote side
truncate value without knowing about UTF-8 symbols.

Signed-off-by: Kirill A. Korinsky <kirill@korins.ky>