Issue #4443 -  Track backport of ALPN APIs to Java 8
For Jetty 9.2.x, supporting the backport of the JDK ALPN APIs
of JDK 8u252 would have required a number of backports from
Jetty 9.4.x.
Rather than doing all that work for an EOL Jetty branch, we
instead excluded the modules/tests that required ALPN, so
that the Jetty 9.2.x branch can still be built with 8u252
or later, but it will lack some of the artifacts.
Users needing those artifacts are encouraged to move to
Jetty 9.4.x or later.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>