Issue #4443 - Track backport of ALPN APIs to Java 8.
Continuous integration builds for Jetty 9.3.x are now done with Java 9.
Jetty 9.3.x needs Java 9+ to be built in order to build also the ALPN
modules that depend on Java 9 APIs.
Jetty 9.3.x would not be buildable with OpenJDK 8u252 or later,
but it is an EOL branch and required Java 9 to be built anyway.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>