Add GracefulShutdownHandler
This adds a handler that takes care of keeping the connection alive
during a shutdown until the request is finished or the graceful stop
timeout is reached.

This is similar to what the StatisticsHandler already does, but does
not collect the metrics, thereby reducing resource usage. Also the
StatisticsHandler does not wait for requests in dispatched state.

Signed-off-by: Thomas Draebing <thomas.draebing@sap.com>