Issue #5442 - allow OpenIdConfigurations to be stored as beans on the server
Multiple OpenIdConfigurations can be stored on server and one is selected
based on the realm name. This allows an option of multiple configurations
where the webapp selects the configuration it wants.

The OpenIdLoginService is no longer needed to be set. It can be null in which
case the Authenticator creates its own one. It can be set to a different
LoginService such as HashLoginService and then the OpenIdAuthenticator will
wrap that LoginService. Otherwise if a OpenIdLoginService is specifically
set it will be used.

Signed-off-by: Lachlan Roberts <lachlan@webtide.com>