Issue #5830 Remove com.sun.nio.file import.
This is acheived by using the 3-arg Class.forName() method call.
The bnd tool which generates the osgi manifests doesn't seem to
include Import-Package statements for the 3-arg method call, but
does for the 1-arg method call. The speculative loading of the
com.sun.nio.file.SensitivityWatchEventModifier will probably be
removed in jetty-10 onwards.

Signed-off-by: Jan Bartel <janb@webtide.com>